+++ module/actor/sheets/base.js
@@ -119,6 +119,17 @@ export class ActorSheetPF extends ActorSheet {
     data.labels = this.actor.labels || {};
     data.filters = this._filters;
 
+    // Add inventory value
+    {
+      const gpValue = this.calculateTotalItemValue();
+      const totalValue = {
+        gp: Math.max(0, Math.floor(gpValue)),
+        sp: Math.max(0, Math.floor(gpValue*10 - Math.floor(gpValue)*10)),
+        cp: Math.max(0, Math.floor(Math.floor(gpValue*100 - Math.floor(gpValue)*100) - (Math.floor(gpValue*10 - Math.floor(gpValue)*10)*10))),
+      };
+      data.labels.totalValue = game.i18n.localize("PF1.ItemContainerTotalItemValue").format(totalValue.gp, totalValue.sp, totalValue.cp);
+    }
+
     // Hit point sources
     if (this.actor.sourceDetails != null) data.sourceDetails = expandObject(this.actor.sourceDetails);
     else data.sourceDetails = null;
@@ -602,6 +613,7 @@ export class ActorSheetPF extends ActorSheet {
 
     // Click to change text input
     html.find('*[data-action="input-text"]').click(event => this._onInputText(event));
+    html.find('*[data-action="input-text"].wheel-change').on("wheel", event => this._onInputText(event.originalEvent));
 
     // Item Dragging
     let handler = ev => this._onDragStart(ev);
@@ -794,10 +806,6 @@ export class ActorSheetPF extends ActorSheet {
     html.find(".item-list .spell-uses input[data-type='max']").off("change")
     .on("change", this._setMaxSpellUses.bind(this))
     .on("wheel", this._setMaxSpellUses.bind(this));
-
-    html.find(".spell-level-uses").off("change")
-    .on("change", this._setSpellbookUses.bind(this))
-    .on("wheel", this._setSpellbookUses.bind(this));
 
     html.find(".spell-points-current .value input[type='text']").off("change")
     .on("change", this._setSpellPoints.bind(this))
@@ -1164,36 +1172,41 @@ export class ActorSheetPF extends ActorSheet {
    */
   _onInputText(event) {
     event.preventDefault();
-    const elem = this.element.find(event.currentTarget.dataset.for);
+    const forStr = event.currentTarget.dataset.for;
+    let elem;
+    if (forStr.match(/CHILD-([0-9]+)/)) {
+      const n = parseInt(RegExp.$1);
+      elem = $(event.currentTarget.children[n]);
+    }
+    else {
+      elem = this.element.find(event.currentTarget.dataset.for);
+    }
+    if (!elem || (elem && elem.attr("disabled"))) return;
 
     const [prevName, prevValue] = [elem.attr("name"), elem.attr("value")];
     elem.prop("readonly", false);
     elem.attr("name", event.currentTarget.dataset.attrName);
     let value = getProperty(this.actor.data, event.currentTarget.dataset.attrName);
     elem.attr("value", value);
-    elem.select();
+
+    const wheelEvent = (event && event instanceof WheelEvent);
+    if (wheelEvent) {
+      this._mouseWheelAdd(event, elem[0]);
+    }
+    else {
+      elem.select();
+    }
 
     const handler = event => {
-      elem[0].removeEventListener("focusout", handler);
+      if (wheelEvent) elem[0].removeEventListener("mouseout", handler);
+      else elem[0].addEventListener("focusout", handler);
       elem[0].removeEventListener("click", handler);
-      if (typeof value === "number") value = value.toString();
-      if (value !== elem.attr("value")) {
-        this._onSubmit(event);
-      }
-      else {
-        window.getSelection().removeAllRanges();
-        if (prevName) {
-          elem.attr("name", prevName);
-        }
-        else {
-          elem.removeAttr("name");
-        }
-        elem.attr("value", prevValue);
-        elem.prop("readonly", true);
-      }
+
+      this._onSubmit(event);
     };
 
-    elem[0].addEventListener("focusout", handler);
+    if (wheelEvent) elem[0].addEventListener("mouseout", handler);
+    else elem[0].addEventListener("focusout", handler);
     elem[0].addEventListener("click", handler);
   }
 
@@ -2034,6 +2047,7 @@ export class ActorSheetPF extends ActorSheet {
       "data.attributes.vigor.temp",
       "data.attributes.wounds.value",
     ];
+
     for (let [k, v] of Object.entries(formData)) {
       if (typeof v !== "string") continue;
       // Add or subtract values
@@ -2071,11 +2085,18 @@ export class ActorSheetPF extends ActorSheet {
     return super._updateObject(event, formData);
   }
 
-  // async _renderInner(data, options) {
-    // let t1 = new Date();
-    // const result = await super._renderInner(data, options);
-    // let t2 = new Date();
-    // console.trace(t2 - t1);
-    // return result;
-  // }
+  calculateTotalItemValue() {
+    const items = this.actor.items.filter(o => o.data.data.price != null);
+    return items.reduce((cur, i) => {
+      return cur + i.getValue({ sellValue: 1 });
+    }, 0);
+  }
+
+  calculateSellItemValue() {
+    const items = this.actor.items.filter(o => o.data.data.price != null);
+    const sellMultiplier = this.actor.getFlag("pf1", "sellMultiplier") || 0.5;
+    return items.reduce((cur, i) => {
+      return cur + i.getValue({ sellValue: sellMultiplier });
+    }, 0);
+  }
 }